requirements.txt
Django==2.1.7
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
pytz==2018.9